requirement.txt
asn1crypto==0.24.0
azure-common==1.1.15
azure-nspkg==2.0.0
azure-storage==0.36.0
boto3==1.8.9
botocore==1.11.9
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
docutils==0.14
future==0.16.0
idna==2.7
ijson==2.3
jmespath==0.9.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycryptodomex==3.6.6
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
snowflake-connector-python==1.6.9
urllib3==1.23
